Rename main.c to programming problem
A certain store keeper needs to automate his sales counter. He plans to store a maximum of 5 items in his store. An item he wants is described by a unique code and a name. He wants that he should be able to discontinue an item and introduce a new one in his store as per the requirement of his customers. An item he has will have a price associated with it and the total units available in his store. Design a system in which he will be able to introduce an item. Specify a price, buy new stock from market and add to his store and sell an item to his customer. A customer may demand any number of units of any item. A customer should also be able to enquire availability and price of an item.

(Introduce new item to the store. Buy new stock and add it to the store. List the available items in the store. Sale an item from his store. Enquire the price of an item. Find out the total worth of his stock. Remove an item from his store)